Extracting a couple Parameter components; Creating util parameterHelper function for validation
